% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_dcpo.R
\docType{data}
\name{cc_dcpo}
\alias{cc_dcpo}
\title{Custom Country Names Used by DCPO}
\format{
A dataframe with six columns
}
\usage{
cc_dcpo
}
\description{
A dataframe of custom country names for use with countrycode::countrycode
}
\details{
\itemize{
  \item country.name.en
  \item country.name.en.regex
  \item iso2c
  \item iso3c
  \item wb_api3c
  \item dcpo.name
}
}
\keyword{datasets}
