% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpo_config.R
\name{dcpo_config}
\alias{dcpo_config}
\title{Configure system files for compiling Stan source code}
\usage{
dcpo_config(pkgdir = ".")
}
\value{
Invisibly, whether or not any files were added/removed/modified by
  the function.
}
\description{
Creates or update package-specific system files to compile `.stan` model
files found in `inst/stan`.
}
\details{
The Stan source files for the package should be stored in:
\itemize{
  \item `inst/stan` for `.stan` files containing instructions to
  build a `stanmodel` object.
  \item `inst/stan/any_subfolder` for files to be included via the
  `#include "/my_subfolder/mylib.stan"` directive.
  \item `inst/stan/any_subfolder` for a `license.stan` file.
  \item `inst/include` for the `stan_meta_header.hpp` file, to be
  used for directly interacting with the Stan C++ libraries.
}
}
