% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_claassen.R
\name{format_claassen}
\alias{format_claassen}
\title{Format DCPO Data for Estimation Using Claassen's (2019) Stan Files}
\usage{
format_claassen(dcpo_data)
}
\arguments{
\item{dcpo_data}{a data frame of survey responses generated by \code{dcpo_setup}}
}
\value{
a list of Stan data
}
\description{
\code{format_claassen} formats DCPO data output by \code{dcpo_setup} for use with Claassen's
 Stan files for "Estimating Smooth Country-Year Panels for Public Opinion" (_Political Analysis_,
 2019).
}
\details{
\code{format_claassen} formats DCPO data output by \code{dcpo_setup} for use with Claassen's
Stan files for "Estimating Smooth Country-Year Panels for Public Opinion" (_Political Analysis_,
2019).  Following Claassen, it dichotomizes variables by coding responses above the midpoint of
the response scale as 1.
}
