% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dcpo.R
\name{format_dcpo}
\alias{format_dcpo}
\title{Format DCPO Data for Estimation Using DCPO Stan Files}
\usage{
format_dcpo(dcpo_data, scale_q, scale_cp, delta = TRUE)
}
\arguments{
\item{dcpo_data}{a data frame of survey responses generated by \code{dcpo_setup}}

\item{scale_q}{a string indicating a survey question in dcpo_data to assign difficulty .5 at its cutpoint scale_cp}

\item{scale_cp}{an integer indicating which cutpoint for scale_q should be assigned difficulty .5}

\item{delta}{logical: should item difficulty be estimated to vary by question-country to account for potential lack of equivalence?}
}
\value{
a list of Stan data
}
\description{
\code{format_dcpo} formats DCPO data output by \code{dcpo_setup} for use with the \code{DCPO}
 package's \code{dcpo} Stan files.
}
