% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dgirt.R
\name{format_dgirt}
\alias{format_dgirt}
\title{Format DCPO Data for Estimation Using DGIRT Stan Files}
\usage{
format_dgirt(dcpo_data)
}
\arguments{
\item{dcpo_data}{a data frame of survey responses generated by \code{dcpo_setup}}
}
\value{
a list of Stan data
}
\description{
\code{format_dgirt} formats DCPO data output by \code{dcpo_setup} for use with the \code{dgo}
 package's \code{dgirt} Stan files.
}
