% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validation_tests.R
\name{internal_validation_tests}
\alias{internal_validation_tests}
\title{Internal Validation Tests for DCPO Models}
\usage{
internal_validation_tests(dcpo_input, dcpo_output)
}
\arguments{
\item{dcpo_input}{the data object created by \code{dcpo_setup} used as input for \code{DCPO::dcpo}}

\item{dcpo_output}{the output of \code{DCPO::dcpo}}
}
\value{
A data frame of validation test statistics, including the model's mean absolute error,
the percentage improvement of this mean absolute error over that of the baseline provided by
the country means in the available survey data, and the leave-one-out information criterion
proposed by Ventari, Gelman, Gabry (2017).  The country-means mean absolute error is also
reported for reference.
}
\description{
Internal Validation Tests for DCPO Models
}
\details{

}
