% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_corr.R
\name{mean_corr}
\alias{mean_corr}
\title{Check the Mean Correlations of Survey Items}
\usage{
mean_corr(dcpo_input_raw)
}
\arguments{
\item{dcpo_input_raw}{a data frame output by \code{dcpo_setup}}
}
\value{
a data frame
}
\description{
\code{mean_corr} summarizes the correlations across items.
}
\details{
\code{mean_corr}, when passed a data frame of survey item
observations, calculates the correlations between all pairs of items
and cutpoints and returns, for each item-cutpoint, its mean correlation
with all other included item-cutpoints.  This can be useful for
identifying data-entry problems as well as possible multidimensionality
issues.
}
