% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_min_cy.R
\name{with_min_cy}
\alias{with_min_cy}
\title{Select Items with Observations in a Minimum Number of Distinct Country-Years}
\usage{
with_min_cy(x, min_yrs)
}
\arguments{
\item{x}{a data frame of survey items and marginals generated by \code{dcpo_setup}}

\item{min_yrs}{the minimum number of distinct country-years needed for an item included in x to be retained}
}
\value{
a data frame
}
\description{
\code{with_min_cy} is a convenience function for filtering data to be input to \code{dcpo} by the number of country-year-items observed
}
