% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dcpo.R
\name{format_dcpo}
\alias{format_dcpo}
\title{Format DCPO Data for Estimation Using DCPO Stan Files}
\usage{
format_dcpo(x, scale_q, scale_cp)
}
\arguments{
\item{dcpo_data}{a data frame of survey responses generated by \code{dcpo_setup}}
}
\value{
a list of Stan data
}
\description{
\code{format_dcpo} formats DCPO data output by \code{dcpo_setup} for use with the \code{DCPO}
 package's \code{dcpo} Stan files.
}
